risk_scoring_api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
scikit-learn>=1.3.0
pydantic>=2.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
